only one instance of each Activity, always in a single Task
This makes it so there is only ever a single instances of any of the
Activities in this app.  It would be confusing if there were multiple
instances of the trigger/swipe or countdown screen active.  Also, these
should all only ever be in a single Task, and nothing else from outside
apps should be in that Task.